More Chores
Added multiple missing drawables and a couple json files to get the app running. Also commented out a MediaPlayer variable called "snap_sound" (in MinfulnessMeditationGame.java) for the meditation game because I couldn' find anything for it at the moment and it doesn't seem required to get the app to build and function.