Addition of Notification to model and accounting for that
The Notification model needed to be added to the phone as it is on the backend of the server. I also started changing the parts where UserInteraction was used incorrectly into Notification. Mainly the collection of the necessary notifications. I will might use UserInteraction to find all the notifications required, or add the receiver and initiator to Notification.